refactor: passando buttonManager para dentro de Logic
