+++ src/games/vowels/scenes/GameScene.ts
@@ -1,15 +1,11 @@
-import ButtonManager from "../../common/managers/ButtonManager";
 import Logic from "../logic/Logic";
 import Phaser from "phaser";
 
 export default class GameScene extends Phaser.Scene {
-  private buttonManager: ButtonManager;
   private logic: Logic;
 
   constructor() {
-    super("Vowels");
-
-    this.buttonManager = new ButtonManager(this);
+    super("vowelsGameScene");
     this.logic = new Logic(this);
   }
 
@@ -23,18 +19,8 @@ export default class GameScene extends Phaser.Scene {
   }
 
   create() {
-    const buttonPositions: { x: number; y: number }[] = [
-      { x: 200, y: 500 },
-      { x: 400, y: 500 },
-      { x: 600, y: 500 },
-    ];
-    const buttonTextures: string[] = [
-      "defaultButton",
-      "hoverButton",
-      "clickedButton",
-    ];
-    this.buttonManager.createButtons(buttonPositions, buttonTextures);
     this.logic.createImage(this.logic.accessCurrentLevel().getName());
+    this.logic.createButtons();
     this.setupLevel();
 
     console.log("Jogo das vogais carregado!");
@@ -43,28 +29,21 @@ export default class GameScene extends Phaser.Scene {
   update() {}
 
   setupLevel() {
-    const answer: string = this.logic.accessCurrentLevel().getAnswer();
-    const buttonsNumber: number = this.buttonManager.getButtons().length;
-    const buttonTexts: string[] = this.logic.generateButtonsLetters(
-      buttonsNumber,
-      answer,
-    );
-    this.buttonManager.setButtonTexts(buttonTexts);
     this.logic.setImageTexture(this.logic.accessCurrentLevel().getName());
+    this.logic.setButtonTexts();
 
-    this.buttonManager.getButtons().forEach((button) => {
+    this.logic.getButtons().forEach((button) => {
       button.off("pointerdown");
       button.on("pointerdown", () => {
-        const result = this.logic.handleAnswer(button, this.time.now, "star");
+        const result = this.logic.handleClick(button, this.time.now);
 
         this.time.delayedCall(1000, () => {
-          if (result.correct) {
+          if (result.correct)
             if (result.finished) {
               this.scene.start("vowelsCredits");
             } else {
-              this.logic.nextLevel(this.buttonManager);
+              this.setupLevel();
             }
-          }
         });
       });
     });
